\name{mosaics-internal}
\title{Internal mosaics objects}
\alias{.adapGridMosaicsZ0_IO}
\alias{.adapGridMosaicsZ0_OS}
\alias{.adapGridMosaicsZ0_TS}
\alias{.calcModelBIC}
\alias{.calcPN}
\alias{.calMDZ1_2S}
\alias{.computeStat}
\alias{.emZ1_2S}
\alias{.eStepZ1_2S}
\alias{.exportBED}
\alias{.exportGFF}
\alias{.exportTXT}
\alias{.getParamZ0}
\alias{.getPH_1S}
\alias{.getPH_2S}
\alias{.getPi0}
\alias{.margDist_1S}
\alias{.margDist_2S}
\alias{.margDistZ1_1S}
\alias{.margDistZ1_2S}
\alias{.mosaicsFit_IO}
\alias{.mosaicsFit_OS}
\alias{.mosaicsFit_TS}
\alias{.mosaicsZ0_IO}
\alias{.mosaicsZ0_OS}
\alias{.mosaicsZ0_TS}
\alias{.mosaicsZ1_1S}
\alias{.mosaicsZ1_2S}
\alias{.mStepZ1_2S}
\alias{.peakCall}
\alias{.plotGOF}
\alias{.processBin_MGC}
\alias{.processBin_MGCX}
\alias{.processBin_X}
\alias{.rlmFit_IO}
\alias{.rlmFit_OS}
\alias{.rlmFit_TS}
\alias{.reportSummary}
\alias{.validDataType}
\alias{.validLocation}
\alias{.validType}
\alias{conv_1S}
\alias{conv_2S}
\description{Internal mosaics objects.}
\details{These are not to be called by the user.}
\keyword{internal}
